Move merkle_error to merkle::error
This error module is only internal to the merkle module, with one of its
structs being exported by the parent merkle module.
