+++ libtransact/src/state/merkle/mod.rs
@@ -16,6 +16,8 @@
  * ------------------------------------------------------------------------------
  */
 
+mod error;
+
 use std::collections::BTreeMap;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::io::Cursor;
@@ -34,7 +36,7 @@ use super::change_log::{ChangeLogEntry, Successor};
 use super::error::{StatePruneError, StateReadError, StateWriteError};
 use super::{Prune, Read, StateChange, Write};
 
-pub use super::merkle_error::StateDatabaseError;
+pub use self::error::StateDatabaseError;
 
 const TOKEN_SIZE: usize = 2;
